Extract user password check on forms into concern
