+++ cosmetics-web/app/forms/secondary_authentication/app/setup_form.rb
@@ -3,20 +3,16 @@ module SecondaryAuthentication
     class SetupForm
       include ActiveModel::Model
       include ActiveModel::Attributes
+      include UserPasswordCheckFormValidation
 
       attribute :app_authentication_code
       attribute :secret_key
-      attribute :password
       attribute :user
 
       validates :app_authentication_code, presence: true
       validate :app_authentication_code, :validate_app_authentication_code
 
       delegate :qr_code, :last_totp_at, to: :secondary_authentication
-
-      validates :password, presence: true
-
-      validate :correct_password
 
       def [](field)
         public_send(field.to_sym)
@@ -40,14 +36,6 @@ module SecondaryAuthentication
 
       def secondary_authentication
         @secondary_authentication ||= SecondaryAuthentication::TimeOtp.new(user, secret_key)
-      end
-
-      def correct_password
-        return if errors[:password].present?
-
-        unless user.valid_password?(password)
-          errors.add(:password, :invalid)
-        end
       end
 
       def validate_app_authentication_code
