+++ cosmetics-web/app/forms/secondary_authentication/sms/setup_form.rb
@@ -4,9 +4,9 @@ module SecondaryAuthentication
       include ActiveModel::Model
       include ActiveModel::Attributes
       include StripWhitespace
+      include UserPasswordCheckFormValidation
 
       attribute :mobile_number
-      attribute :password
       attribute :user
 
       validates :mobile_number, presence: { message: :invalid }
@@ -17,19 +17,7 @@ module SecondaryAuthentication
                 phone: { message: :invalid, allow_international: false },
                 if: :validate_uk_mobile_number?
 
-      validates :password, presence: true
-
-      validate :correct_password
-
     private
-
-      def correct_password
-        return if errors[:password].present?
-
-        unless user.valid_password?(password)
-          errors.add(:password, :invalid)
-        end
-      end
 
       def validate_uk_mobile_number?
         mobile_number.present? && user && !user.class::ALLOW_INTERNATIONAL_PHONE_NUMBER
